requirements.txt
# KPI Pipeline Requirements
# Python 3.9+

# Core data processing
pandas>=2.0.0

# Configuration management
pyyaml>=6.0

# Excel export
openpyxl>=3.1.0

# Date/time handling
python-dateutil>=2.8.0
